inference defined in inference not __init__
